refactor(#451): split up controllers (wip)
